% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.R
\name{check_fit}
\alias{check_fit}
\title{A function to test how well the fit matches the data}
\usage{
check_fit(fit_output, dist, ps)
}
\arguments{
\item{fit_output}{A data frame output from mix_dist()}

\item{dist}{A numeric vector defining the distribution}

\item{ps}{A numeric vector describing the granule sizes}
}
\value{
a ggplot object
}
\description{
A function to test how well the fit matches the data
}
