% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dist}
\alias{Dist}
\title{An distribution data set}
\format{A data frame with 100 rows and 4 variables:
\describe{
\item{size}{Size of granule, in um}
\item{118 - Average}{Distribution of sample}
\item{206 - Average}{Distribution of sample}
\item{85 - Average}{Distribution of sample}
}}
\usage{
Dist
}
\description{
A dataset containing the distributions from three samples
}
\keyword{datasets}

