% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.R
\name{get_heights}
\alias{get_heights}
\title{Get the heights at peaks}
\usage{
get_heights(dist, ps, means)
}
\arguments{
\item{dist}{A numeric vector defining the distribution}

\item{ps}{A numeric vector describing the granule sizes}

\item{means}{A named numeric vector defining the means (center) for each peak}
}
\value{
A named vector with heights for each mean
}
\description{
Get the heights at peaks
}

