% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular_plots.R
\name{ggfit}
\alias{ggfit}
\title{A function for plotting a fit output against the true distribution}
\usage{
ggfit(fit_output, dist, ps)
}
\arguments{
\item{fit_output}{A data frame output from mix_dist()}

\item{dist}{A numeric vector defining the distribution}

\item{ps}{A numeric vector describing the granule sizes}
}
\value{
a ggplot object
}
\description{
A function for plotting a fit output against the true distribution
}

