% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular_plots.R
\name{ggfit_grp_tbl}
\alias{ggfit_grp_tbl}
\title{Generate fit plots for a grouped tbl}
\usage{
ggfit_grp_tbl(.data, fit_output, proportion, size)
}
\arguments{
\item{.data}{A tbl output from mix_grp_tbl, with list columns}

\item{fit_output}{Variable name holding mix_dist_output}

\item{proportion}{Variable name holding distribution as a list}

\item{size}{Variable name holding particle size as a list}
}
\value{
The original tbl with a list column of ggplot output
}
\description{
Generate fit plots for a grouped tbl
}
