% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular_plots.R
\name{make_dist}
\alias{make_dist}
\title{Make a mixture distribution}
\usage{
make_dist(fit_output, ps)
}
\arguments{
\item{fit_output}{A data frame output from mix_dist()}

\item{ps}{A numeric vector describing the granule sizes}
}
\value{
A numeric vector describing the mixture distribution
}
\description{
Make a mixture distribution
}
