% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.R
\name{mix_dist_df}
\alias{mix_dist_df}
\title{Get just the data.frame output from mix_dist}
\usage{
mix_dist_df(dist_df, ps, comp_means, dist_name, printFit = TRUE,
  printPlot = TRUE, emnum = 5)
}
\arguments{
\item{dist_df}{A data.frame containing the distribution}

\item{ps}{A numeric vector describing the granule sizes}

\item{comp_means}{A named numeric vector defining the means (center) for each peak}

\item{dist_name}{A string defining the name of the distribution}

\item{printFit}{Logical. Whether or not to print the fir output to the console}

\item{printPlot}{Logical. Whether or not to print the plot showing the fit}

\item{emnum}{passed to mix() - A non-negative integer specifying the number of EM steps to be performed}
}
\value{
A data.frame with the fit parameters for each distribution
}
\description{
Get just the data.frame output from mix_dist
}
