% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.R
\name{mix_grp_tbl}
\alias{mix_grp_tbl}
\title{Title}
\usage{
mix_grp_tbl(.data, proportion, size, comp_means)
}
\arguments{
\item{.data}{A tbl grouped by each distribution}

\item{proportion}{An unquoted variable name}

\item{size}{An unquoted variable name}

\item{comp_means}{A named vector defining the means of each component}
}
\value{
A mutated tbl with list column output
}
\description{
Title
}
