% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms2}
\alias{ms2}
\title{An example mix_dist output}
\format{A data frame with 3 rows and 8 variables:
\describe{
\item{sample}{Name of the sample}
\item{peak}{Peak name}
\item{pi}{Distribution proportion}
\item{mu}{Distribution mean}
\item{sigma}{Distribution sd}
\item{pi.se}{proportion standard error}
\item{mu.se}{mean standard error}
\item{sigma.se}{sd standard error}
}}
\usage{
ms2
}
\description{
A dataset containing the output from 85 - Average
}
\keyword{datasets}

