% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_granular}
\alias{run_granular}
\title{Run the granular shiny app locally}
\usage{
run_granular(port = NULL, launch.browser = getOption("shiny.launch.browser",
  interactive()), host = getOption("shiny.host", "127.0.0.1"))
}
\arguments{
\item{port}{The TCP port that the application should listen on. Defaults to choosing a random port.}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started. defaults to true in interactive sessions only.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not.}
}
\description{
Run the granular shiny app locally
}

