% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular_plots.R
\name{ggfit_grp_tbl}
\alias{ggfit_grp_tbl}
\title{Generate fit plots for a grouped tbl}
\usage{
ggfit_grp_tbl(.data, fit_output, dist, ps)
}
\arguments{
\item{.data}{A grouped tbl}

\item{fit_output}{Output from mix_dist}

\item{dist}{A vector describing the distribution}

\item{ps}{A vector describing the particle sizes}
}
\value{
A tbl with a list column of ggplot output
}
\description{
Generate fit plots for a grouped tbl
}
