% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granular.R
\name{mix_dist}
\alias{mix_dist}
\title{use mix() to estimate underlying distributions}
\usage{
mix_dist(dist, ps, comp_means, sample_name = NULL, emnum = 10)
}
\arguments{
\item{dist}{A numeric vector defining the distribution}

\item{ps}{A numeric vector describing the granule sizes}

\item{comp_means}{A named numeric vector defining the means (center) for each peak}

\item{sample_name}{An optional name for the sample}

\item{emnum}{passed to mix() - A non-negative integer specifying the number of EM steps to be performed}
}
\value{
A list with the fit parameters for each distribution, and complete output from mixdist::mix()
}
\description{
use mix() to estimate underlying distributions
}
